Make Plots an optional dependency
